+++ tests/acceptance/logged-out-homepage-test.ts
@@ -11,14 +11,14 @@ import { click, setupOSFApplicationTest } from 'ember-osf-web/tests/helpers';
 
 const { featureFlagNames: { ABTesting } } = config;
 
-module('Acceptance | new home page test', hooks => {
+module('Acceptance | logged-out home page test', hooks => {
     setupOSFApplicationTest(hooks);
     setupMirage(hooks);
 
-    test('visiting new-home', async assert => {
-        await visit('/new-home');
+    test('visiting home', async assert => {
+        await visit('/');
 
-        assert.equal(currentURL(), '/new-home', "Still at 'new-home'.");
+        assert.equal(currentURL(), '/', "Still at 'home'.");
         // Check navbar
         assert.dom('nav.navbar').exists();
         assert.dom('nav.navbar .sign-in').exists();
@@ -39,14 +39,45 @@ module('Acceptance | new home page test', hooks => {
         assert.dom('[data-test-arrow]').exists({ count: 3 });
         assert.dom('[data-test-learn-more-button]').exists();
 
-        // Check footer.
+        // Check carousel
+        assert.dom('[data-test-testimonials-container]').exists();
+        assert.dom('[data-test-testimonials-heading]').hasText('What others are saying');
+        assert.dom('[data-test-testimonials-container]').exists();
+        // check for 3 carousel slides
+        assert.dom('[data-test-testimonials-slide-1]').exists();
+
+        await percySnapshot(assert);
+
+        await click('[data-test-carousel-button-next]');
+
+        await percySnapshot(assert);
+
+        assert.dom('[data-test-testimonials-slide-2]').exists();
+        await click('[data-test-carousel-button-next]');
+        assert.dom('[data-test-testimonials-slide-3]').exists();
+
+        // Check integrations
+        assert.dom('[data-test-integrations-heading]')
+            .containsText(t('new-home.integrations-section.header').toString());
+        assert.dom('[data-test-authentication-heading]')
+            .containsText(t('new-home.integrations-section.authentication').toString());
+        assert.dom('[data-test-discovery-heading]')
+            .containsText(t('new-home.integrations-section.discovery').toString());
+        assert.dom('[data-test-references-heading]')
+            .containsText(t('new-home.integrations-section.references').toString());
+        assert.dom('[data-test-storage-heading]')
+            .containsText(t('new-home.integrations-section.storage').toString());
+        assert.dom('[data-test-logo]').exists({ count: 8 });
+        assert.dom('[data-test-get-started-button]').exists({ count: 2 });
+
+        // Check footer
         assert.dom('footer').exists();
         await a11yAudit();
         await percySnapshot(assert);
     });
 
-    test('visiting new-home version B', async function(assert) {
-        await visit('/new-home');
+    test('visiting home version B', async function(assert) {
+        await visit('/');
         const features = this.owner.lookup('service:features') as Features;
 
         features.enable(ABTesting.homePageVersionB);
@@ -55,13 +86,14 @@ module('Acceptance | new home page test', hooks => {
         assert.dom('[data-test-add-research-heading]').exists();
         assert.dom('[data-test-add-research-subheading]').exists();
         assert.ok(document.querySelector('[data-test-hero-container]')!.className.includes('versionB'));
+        assert.dom('[data-test-get-started-button]').exists({ count: 1 });
 
         await a11yAudit();
         await percySnapshot(assert);
     });
 
     test('Get Started button works', async assert => {
-        await visit('/new-home');
+        await visit('/');
 
         await click('[data-test-get-started-button]');
         assert.equal(currentURL(), '/register');